pyproject.toml
[tool.poetry]
name = "hm13-mail-2"
version = "0.1.0"
description = ""
authors = ["Anatolii6974 <anatolii.perfilov@gmail.com>"]
readme = "README.md"
packages = [{include = "hm13_mail_2"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
sqlalchemy = "^2.0.25"
psycopg2 = "^2.9.9"
alembic = "^1.13.1"
libgravatar = "^1.0.4"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-multipart = "^0.0.6"
pydantic = {extras = ["email"], version = "^2.5.3"}
fastapi-mail = "^1.4.1"
bcrypt = "3.2.2"
passlib = "^1.7.4"
python-dotenv = "^1.0.1"
asyncio-redis = "^0.16.0"
fastapi-limiter = "^0.1.6"
cloudinary = "^1.38.0"
pytest = "^8.0.0"


[tool.poetry.group.dev.dependencies]
sphinx = "^7.2.6"


[tool.poetry.group.test.dependencies]
httpx = "^0.26.0"

[tool.pytest.ini_options]
pythonpath = ["."]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
